+++ mlxtend/image/tests/test_extract_face_landmarks.py
@@ -43,31 +43,24 @@ def test_jpg():
 
     assert landmarks2.shape == (68, 2)
 
+    true_vals = np.array([[85, 110],
+                          [85, 120],
+                          [85, 130],
+                          [87, 140],
+                          [91, 148],
+                          [97, 155],
+                          [104, 160],
+                          [112, 164],
+                          [120, 165],
+                          [125, 162]], dtype=np.int32)
+
     if os.name == 'nt':
-        true_vals = np.array([[85, 111],
-                              [84, 120],
-                              [85, 130],
-                              [87, 140],
-                              [91, 148],
-                              [98, 155],
-                              [105, 160],
-                              [113, 164],
-                              [120, 165],
-                              [126, 162]], dtype=np.int32)
-
+        # on windows, imageio parses jpgs sometimes differently so pixel values
+        # maybe slightly different
+        assert np.sum(np.abs(landmarks2[:10] - true_vals) > 2) == 0
     else:
-        true_vals = np.array([[85, 110],
-                              [85, 120],
-                              [85, 130],
-                              [87, 140],
-                              [91, 148],
-                              [97, 155],
-                              [104, 160],
-                              [112, 164],
-                              [120, 165],
-                              [125, 162]], dtype=np.int32)
-
-    np.testing.assert_array_equal(landmarks2[:10], true_vals)
+        assert np.sum(np.abs(landmarks2[:10] - true_vals) > 0) == 0
+        np.testing.assert_array_equal(landmarks2[:10], true_vals)
 
 
 def test_grayscale():
@@ -95,5 +88,4 @@ def test_noface():
     img = imageio.core.util.Array((
       np.random.random((193, 341, 3))*255).astype(np.uint8))
     landmarks = extract_face_landmarks(img)
-    assert landmarks.shape == (68, 2)
-    np.testing.assert_array_equal(landmarks, np.zeros((68, 2)))
+    assert landmarks is None